refactor(products): remove GetAllProducts and replace with GetAllProductCards
- Removed legacy GetAllProductsHandler and query
- Introduced GetAllProductCards as a more accurate representation for card view
- Cleaned up folder and file naming for consistency

+++ AmazonKiller.Application/DTOs/Account/Wishlist/WishlistItemDto.cs
@@ -6,5 +6,5 @@ public record WishlistItemDto(
     string ImageUrl,
     decimal Price,
     decimal? OldPrice,
-    double Rating
+    decimal Rating
 );
